implemented some fixes and moved enums to seperate file
