+++ scripts/events_module/patrol/patrol_outcome.py
@@ -25,7 +25,8 @@ from scripts.utility import (
 )
 from scripts.game_structure.game_essentials import game
 from scripts.cat.skills import SkillPath
-from scripts.cat.cats import Cat, ILLNESSES, INJURIES, PERMANENT, CatAgeEnum
+from scripts.cat.cats import Cat, ILLNESSES, INJURIES, PERMANENT
+from scripts.cat.enums import CatAgeEnum
 from scripts.cat.pelts import Pelt
 from scripts.cat_relations.relationship import Relationship
 from scripts.clan_resources.freshkill import (
