+++ scripts/events_module/relationship/pregnancy_events.py
@@ -3,7 +3,8 @@ from random import choice, randint
 
 import ujson
 
-from scripts.cat.cats import Cat, CatAgeEnum
+from scripts.cat.cats import Cat
+from scripts.cat.enums import CatAgeEnum
 from scripts.cat.history import History
 from scripts.cat.names import names, Name
 from scripts.cat_relations.relationship import Relationship
